Module core_protocol::ticket_aggregation::processor

Structs

AwaitingAggregator

TicketAggregationActions

External API for feeding Ticket Aggregation actions into the Ticket Aggregation processor processing the elements independently in the background.

TicketAggregationAwaiter

TicketAggregationFinalizer

TicketAggregationInteraction

Sets up processing of ticket aggregation interactions and returns relevant read and write mechanism.

Enums

TicketAggregationProcessed

Emitted by the processor background pipeline once processed

TicketAggregationToProcess

The input to the processor background pipeline

Constants

TICKET_AGGREGATION_RX_QUEUE_SIZE

TICKET_AGGREGATION_TX_QUEUE_SIZE

Traits

TicketAggregatorTrait

Type Aliases

BasicTicketAggregationActions